+++ src/windows/Tools.cpp
@@ -1451,7 +1451,7 @@ static bool GetProxyUrlFromIE(UnicodeString & Proxy)
   return Result;
 }
 //---------------------------------------------------------------------------
-bool AutodetectProxy(UnicodeString & HostName, int & PortNumber)
+bool AutodetectProxy(UnicodeString & HostName, intptr_t& PortNumber)
 {
   bool Result = false;
